refactor: isolate video conditioning preparation and cleanup (#7124)
